Trigger Rules in Action!!!
New: Now if there is an new R(x, y) assertion entered to the model,
existing
and already visited ForAll and AtMost Concepts are re-triggered. In the
first case the appropriate C(y) assertion is added to the model and in
the second case the same AtMost concept is added to the model in order
to force our model to re-check it in a later moment. Only drawback is
that more then one same AtMost concept may be added where just one will
be sufficient, for example if there are two subsequent R(x,y), R(x,y)
assertions.

New: The TList.append() method was enriched so that we can add an
assertion to the model without checking for duplicates. This was
necessary for the AtMost re-addition to the same model.

New: The TList.duplicate() method no is using the top-down approach,
i.e., we build the copy from the leaf to the root, in order to keep the
most recent AtMost duplicate (possible multiple AtMost concepts due to
the trigger rule functionality)